requirements.txt
requests
python-whois
python-dotenv
Pillow  # For handling images in Tkinter
